requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
passlib==1.7.4
python-dotenv==1.0.0
Werkzeug==2.3.7
razorpay==1.4.1
